Make AddressBus not self-referring
The AddressBus models, as the name suggests, the address bus, and not RAM

The RAM is then connected to the address bus (and data bus), so that it can be
accessed.

It is however quite common to add a demux on the upper bits in the address bus,
splitting out the address bus in several busses with shorter word length, and
different chip enable/output enable signals.

One way of implementing that, is to have an object which demux the address bus
into several other address busses. But that would need to refer to the smaller
busses as trait objects (Box<dyn AddressBus>)

Since the AddressBus implements copy_from(&mut self, other: &Self), which refer
to Self as a type, it can't be used as a trait object. Also the copy_from
indicates that "other" actually stores values that can be copied, which might
not be true if the peripheral would be something like an I/O or random number
generator.

The proper way to implement a copy_from() if necessary would be to have it as
another trait, for example CopyableMemory. But since copy_from currently only
is used for testing, it makes sense to keep it simple, and add it directly to
the explicit types.

Note that a trait object creates overhead for memory access. So to be able to
have a fast cache/ram hard coded, it still makes sense to have ConfiguredCore
as a generic over the AddressBus